filename generated from groupname plus topic plus date
